requirements.txt
# Data Manipulation
pandas==2.2.3
numpy==2.1.3

# Database Connectivity (The Bridge)
sqlalchemy==2.0.39
psycopg2-binary==2.9.11

# API & Web Scraping
requests==2.32.3

# Modeling & Math
scikit-learn==1.6.1
statsmodels==0.14.4

# Visualization
matplotlib==3.10.0
seaborn==0.13.2

# Utilities
python-dotenv>=1.0.0   # To hide your database passwords (security best practice)
jupyter         # To run the notebooks

# Unit Testing
pytest>=7.0.0          # To run unit tests

# Code Quality
black>=23.12.0 
pylint>=2.17.0 